+++ packages/core/src/controllers/types.ts
@@ -108,6 +108,16 @@ export interface MfaPerformActionOptions {
    * This typically contains user input or challenge-response data.
    */
   payload?: string;
+
+  /**
+   * Enables passkey support in browser autofill suggestions (conditional UI), if supported.
+   */
+  autoFill?: boolean;
+
+  /**
+   * A human-palatable name for the user account, intended only for display on your passkeys..
+   */
+  displayName?: string;
 }
 
 /**